Basic Dashboard Functionality Developed

+++ backend/app/crud/transaction.py
@@ -1,15 +1,16 @@
 from sqlalchemy.orm import Session
 from app.models.transaction import Transaction
 from app.schemas.transaction import TransactionCreate, TransactionUpdate
-from datetime import date
+from datetime import datetime
 from sqlalchemy.orm import joinedload
+from decimal import Decimal
 
 def create_transaction(db: Session, transaction: TransactionCreate):
     db_transaction = Transaction(
         user_id=transaction.user_id,
         category_id=transaction.category_id,
         transaction_name=transaction.transaction_name,
-        amount=float(transaction.amount),
+        amount=Decimal(transaction.amount),
         transaction_date=transaction.transaction_date,
         notes=transaction.notes
     )
@@ -28,14 +29,6 @@ def get_transactions_for_user(db: Session, user_id: int):
 
 def get_transaction(db: Session, transaction_id: int):
     return db.query(Transaction).filter(Transaction.transaction_id == transaction_id).first()
-
-# def get_existing_budget(db: Session, user_id: int, category_id: int, start_date: date, end_date: date):
-#     return db.query(Budget).filter(
-#         Budget.user_id == user_id,
-#         Budget.category_id == category_id,
-#         Budget.start_date == start_date,
-#         Budget.end_date == end_date
-#     ).first()
 
 def delete_transaction(db: Session, transaction_id: int):
     transaction = db.query(Transaction).filter(Transaction.transaction_id == transaction_id).first()
